feat: update insights screen layout with dedicated session details screen
